Implemented Rating slider for single creation version

+++ src/NameSayer/Controller.java
@@ -8,13 +8,8 @@ import javafx.fxml.FXML;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.CheckBoxTableCell;
 import javafx.scene.media.AudioClip;
-import javafx.concurrent.Task;
 import javafx.scene.text.Text;
-
 import java.io.*;
-
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.*;
 
 public class Controller {
@@ -59,15 +54,22 @@ public class Controller {
 			// Display rating if text file exists for that creation name
 			File folder = new File("./Ratings");
 			File[] listOfFiles = folder.listFiles();
+			boolean found = false;
 			for (File file : listOfFiles) {
 				if (file.isFile()) {
 					if (selectedName.equals(file.getName())) {
-
-					}
-					else {
-						wordRating.setText("");
+						found = true;
+                       try {
+                           BufferedReader br = new BufferedReader(new FileReader(file));
+                           wordRating.setText(br.readLine());
+                       } catch (IOException e) {
+                           e.printStackTrace();
+                       }
 					}
 				}
+			}
+			if (found == false) {
+				wordRating.setText("No Rating");
 			}
 
 			if (newValue != null) { // new selection can be null if deleted the last creation
@@ -98,11 +100,8 @@ public class Controller {
 		Playlist.setCellValueFactory(creationData -> {
 			Creation creation = creationData.getValue();
 
-			if (creation.isChecked()) {
-				currentPlaylist.add(creation);
-			} else {
-				currentPlaylist.remove(creation);
-			}
+			if (creation.isChecked()) currentPlaylist.add(creation);
+			else currentPlaylist.remove(creation);
 
 			// when starting a playlist, automatically select the first creation in the queue.
 			if (currentPlaylist.size() == 1) Creations.getSelectionModel().select(currentPlaylist.first());
@@ -124,7 +123,7 @@ public class Controller {
 					Map<String, AudioClip> creationVideos = new HashMap<>();
 
 					// pass in a File object URI string as the Media object does not accept relative paths
-					File[] creationVideoFiles = new File(getCreationDirectory(creation)).listFiles();
+					File[] creationVideoFiles = new File("creations/" + creationName).listFiles();
 
 					if (creationVideoFiles != null) {
 						for (File creationVideoFile : creationVideoFiles) {
@@ -138,16 +137,13 @@ public class Controller {
 		});
 
 		File creationPath = new File("creations");
-		File[] creations = creationPath.listFiles();
-
-		// Create the creation directory  if one doesn't already exist
-		if (!creationPath.exists()) creationPath.mkdir();
-
-		// Populate table with existing Creations and load media files
-		if (creations != null) {
-			for (File creationFolder : creations) {
-				String creationName = creationFolder.getName();
-				data.add(new Creation(creationName));
+		if (creationPath.exists() || creationPath.mkdir()) { // Create the creation directory  if one doesn't already exist
+			File[] creations = creationPath.listFiles();
+			if (creations != null) { // Populate table with existing Creations and load media files
+				for (File creationFolder : creations) {
+					String creationName = creationFolder.getName();
+					data.add(new Creation(creationName));
+				}
 			}
 		}
 
@@ -156,7 +152,7 @@ public class Controller {
 			Creation currentSelection = Creations.getSelectionModel().getSelectedItem();
 			String currentVersion = Versions.getSelectionModel().getSelectedItem();
 
-			if (currentSelection != null) creationPlayers.get(currentSelection.getName()).get(currentVersion).play();
+			if (currentSelection != null && currentVersion != null) creationPlayers.get(currentSelection.getName()).get(currentVersion).play();
 		});
 
 		// Loads the next queued creation. Cycles to the beginning at the end.
@@ -165,11 +161,8 @@ public class Controller {
 			if (currentSelection != null) {
 				Creation nextPlaying = currentPlaylist.higher(currentSelection);
 
-				if (nextPlaying != null) {
-					Creations.getSelectionModel().select(nextPlaying);
-				} else if (currentPlaylist.size() > 0){
-					Creations.getSelectionModel().select(currentPlaylist.first());
-				}
+				if (nextPlaying != null) Creations.getSelectionModel().select(nextPlaying);
+				else if (currentPlaylist.size() > 0) Creations.getSelectionModel().select(currentPlaylist.first());
 			}
 		});
 
@@ -179,33 +172,30 @@ public class Controller {
 			if (currentSelection != null) {
 				Creation nextPlaying = currentPlaylist.lower(currentSelection);
 
-				if (nextPlaying != null) {
-					Creations.getSelectionModel().select(nextPlaying);
-				} else if (currentPlaylist.size() > 0) {
-					Creations.getSelectionModel().select(currentPlaylist.last());
-				}
+				if (nextPlaying != null) Creations.getSelectionModel().select(nextPlaying);
+				else if (currentPlaylist.size() > 0) Creations.getSelectionModel().select(currentPlaylist.last());
 			}
 		});
 
-		// select a random creation from the database and choose the next checked creation
+		// select a random creation from the playlist
 		randomCreation.setOnAction(event -> {
 			Creation currentSelection = Creations.getSelectionModel().getSelectedItem();
+			Creation nextPlaying = currentSelection;
 			if (currentSelection != null) {
-				Creation nextPlaying = currentPlaylist.higher(data.get(random.nextInt(data.size())));
+				if (currentPlaylist.size() > 1) {
+					Iterator<Creation> iteratePlaylist;
+					int randomIndex;
 
-				// ensures that the next creation is not the same one that is currently selected.
-				// if it is, then select the first checked creation after the one that is currently selected.
-				if (nextPlaying != null) {
-					if (nextPlaying.equals(currentSelection)) nextPlaying = currentPlaylist.higher(nextPlaying);
-				} else if (currentPlaylist.size() > 0){
-					if (currentSelection.equals(currentPlaylist.first())) nextPlaying = currentPlaylist.higher(currentPlaylist.first());
-				}
+					// ensures that the next creation is not the same one that is currently selected.
+					// if it is, select a different random creation from the playlist
+					while (nextPlaying.equals(currentSelection)) {
+						randomIndex = random.nextInt(currentPlaylist.size());
+						iteratePlaylist = currentPlaylist.iterator();
+						for (int i = 0; i <= randomIndex; i++) nextPlaying = iteratePlaylist.next();
+					}
 
-				if (nextPlaying != null) {
-					Creations.getSelectionModel().select(nextPlaying);
-				} else if (currentPlaylist.size() > 0){
-					Creations.getSelectionModel().select(currentPlaylist.first());
-				}
+				} else if (currentPlaylist.size() == 1) nextPlaying = currentPlaylist.first();
+				Creations.getSelectionModel().select(nextPlaying);
 			}
 		});
 
@@ -221,18 +211,12 @@ public class Controller {
 				confirmDelete.setContentText("Are you sure you want to delete " + creationName + "?");
 
 				confirmDelete.showAndWait().filter(response -> response == ButtonType.OK).ifPresent(response -> {
-					File creationVersionsDirectory = new File(getCreationDirectory(creation));
+					File creationVersionsDirectory = new File("creations/" + creationName);
 					File[] creationVersions = creationVersionsDirectory.listFiles();
 
-					if (creationVersions != null) {
-						for (File version : creationVersions) {
-							version.delete();
-						}
-					}
-
-					// remove the folder, row, and associated MediaPlayer
-					creationVersionsDirectory.delete();
-					data.remove(creation);
+					// remove the folder, row, and associated players
+					if (creationVersions != null) for (File version : creationVersions) version.delete();
+					if (creationVersionsDirectory.delete()) data.remove(creation);
 				});
 			}
 		});
@@ -254,13 +238,11 @@ public class Controller {
 					confirmOverwrite.setContentText("Would you like to overwrite the existing creation?");
 					confirmOverwrite.showAndWait().filter(response -> response == ButtonType.OK).ifPresent(response -> {
 						data.remove(newCreation);
-						addCreation(newCreation);
+						data.add(newCreation);
 					});
 				} else {
-					File newCreationDirectory = new File(getCreationDirectory(newCreation));
-					newCreationDirectory.mkdir();
-
-					addCreation(newCreation);
+					File newCreationDirectory = new File("creations/" + creationName);
+					if (newCreationDirectory.mkdir()) data.add(newCreation);
 				}
 			} else {
 				Alert invalidCharacters = new Alert(Alert.AlertType.WARNING);
@@ -268,85 +250,7 @@ public class Controller {
 				invalidCharacters.setContentText("Please choose another name for your creation.");
 				invalidCharacters.showAndWait();
 			}
-
 		});
-	}
-
-	/**
-	 * This method uses ffmpeg to create a video and record audio input to add a creation.
-	 * A creation may be redone if the user is not satisfied with the audio.
-	 */
-	private void addCreation(Creation creation) {
-		String creationName = creation.getName();
-		String creationDirectory = getCreationDirectory(creation);
-
-		ProcessBuilder createVideo = new ProcessBuilder("/bin/bash", "-c", "ffmpeg -loglevel quiet -f lavfi -y -i color=c=black:s=320x240:d=5 -vf drawtext=\"fontsize=16:fontcolor=white:x=(w-text_w)/2:y=(h-text_h)/2:text='" + creationName + "'\" \"" + creationDirectory + creationName + "Temp.mp4\"");
-		ProcessBuilder createAudio = new ProcessBuilder("/bin/bash", "-c", "ffmpeg -loglevel quiet -f alsa -y -ac 1 -ar 44100 -i default -t 5 -strict -2 \"" + creationDirectory + creationName + ".m4a\"");
-		ProcessBuilder merge = new ProcessBuilder("/bin/bash", "-c", "ffmpeg -loglevel quiet -y -i \"" + creationDirectory + creationName + "Temp.mp4\" -i \"" + creationDirectory + creationName + ".m4a\" -c copy \"" + creationDirectory + creationName + ".mp4\"");
-
-		Alert voicePrompt = new Alert(Alert.AlertType.INFORMATION);
-		voicePrompt.setHeaderText("Your voice will now be recorded.");
-		voicePrompt.setContentText("Press ENTER or hit OK when you are ready.");
-		voicePrompt.showAndWait();
-
-		Alert voiceRecording = new Alert(Alert.AlertType.INFORMATION);
-		voiceRecording.setHeaderText("Your voice is being recorded.");
-		voiceRecording.setContentText("Please speak clearly into the microphone.");
-		voiceRecording.show();
-
-		Task<Void> recordAudio = new Task<Void>() {
-			@Override
-			protected Void call() throws IOException, InterruptedException {
-				Process createVideoProcess = createVideo.start();
-				createVideoProcess.waitFor();
-				Process createAudioProcess = createAudio.start();
-				createAudioProcess.waitFor();
-				Process mergeProcess = merge.start();
-				mergeProcess.waitFor();
-				return null;
-			}
-		};
-
-		// Adds an event listener when the audio finishes recording.
-		recordAudio.setOnSucceeded(event -> {
-			voiceRecording.close();
-			data.add(creation);
-
-			File rawVideo = new File(creationDirectory + creationName + "Temp.mp4");
-			File rawAudio = new File(creationDirectory + creationName + ".m4a");
-			rawVideo.delete();
-			rawAudio.delete();
-
-			Alert playRecording = new Alert(Alert.AlertType.CONFIRMATION);
-			playRecording.setHeaderText("Your creation has been saved.");
-			playRecording.setContentText("Would you like to hear the recorded audio?");
-			playRecording.showAndWait().filter(response -> response == ButtonType.OK).ifPresent(response -> {
-				Creations.getSelectionModel().select(creation);
-				creationPlayers.get(creationName).get(Versions.getSelectionModel().getSelectedItem()).play();
-
-				Alert keepAudio = new Alert(Alert.AlertType.CONFIRMATION);
-				keepAudio.setHeaderText("Your recording is now playing.");
-				keepAudio.setContentText("Would you like to keep or redo the recording?");
-
-				ButtonType redo = new ButtonType("Redo");
-				ButtonType keep = new ButtonType("Keep");
-
-				keepAudio.getButtonTypes().setAll(keep, redo);
-
-				keepAudio.showAndWait().filter(audioResponse -> audioResponse == redo).ifPresent(audioResponse -> {
-					data.remove(creation);
-					addCreation(creation);
-				});
-			});
-		});
-
-		Thread thread = new Thread(recordAudio);
-		thread.start();
-	}
-
-	private String getCreationDirectory(Creation creation) {
-		String creationName = creation.getName();
-		return "creations/" + creationName + "/";
 	}
 
 	@FXML
@@ -357,7 +261,7 @@ public class Controller {
 	@FXML private void ratingSliderAction() {
 	    if (ratingSlider.getValue() > 50) {
 	    	wordRating.setText("Good");
-	    	try (Writer writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream("filename"), "utf-8"))) {
+	    	try (Writer writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(selectedName), "utf-8"))) {
 	    		writer.write("Good");
 			} catch (IOException e) {
 	    		e.printStackTrace();
@@ -367,7 +271,7 @@ public class Controller {
 		}
 		else {
 			wordRating.setText("Bad");
-			try (Writer writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream("filename"), "utf-8"))) {
+			try (Writer writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(selectedName), "utf-8"))) {
 				writer.write("Bad");
 			} catch (IOException e) {
 				e.printStackTrace();
@@ -383,7 +287,7 @@ public class Controller {
 
 		for (File file : listOfFiles) {
 			if (file.isFile()) {
-				file.renameTo(new File("./Ratings/fileName"));
+				file.renameTo(new File("./Ratings/" + selectedName));
 			}
 		}
 	}
